refactor: rename `authentication` module to `auth`
